ASS/booking_service/requirements.txt
flask
psycopg2-binary
pika
python-dotenv